requirements.txt
streamlit
pypdf2
python-dotenv
faiss-cpu
huggingface_hub
